Login and Register Barebones Completed
MAKE SURE TO READ THIS AND THE CODE ^^^^^^^^^.  There was a very big error in installing and dropping the db. I fixed it (Saahil's db is kinda incorrect. I switched it back to cjs). Login and Register work perfectly. However we might need to make the site for secure with some more validation. A minor css change. Made profiles clickable with a description. Added some QOL changes and debugged some of our sql bugs. Make a account and use the site it works!