Cleaning changes
I swapped files names so that uploadsublewit is the backend for uploading sublewits and sublewit.php is now the frontend page instead of blewit/selectedsublewit. I added comments to each page for dev and reversed saahils change of changing the db id name for the sublewits. (it made it more confusing for the existing code and honeslty didnt fit well). Overall, I did some simple front and backend changes. We are practically done with the project. We just need to discuss the COMMUNITIES BUTTON AND FRONTEND VISUALS